added support for dumping serialized message
The current CLI does not support dumping a serialized transaction, which is a problem for stake pool operators that use multisig wallets. This PR adds a new CLI option, --dump-transaction-message-only, that will instead of sending the transaction just print the transaction to stdout.

By setting that option, it also becomes possible to pass a pubkey to the CLI arguments that normally expect a keypair for signing.